Handle USER in the RUN command
