+++ bin/makisu/cmd/build.go
@@ -69,9 +69,9 @@ type buildCmd struct {
 func getBuildCmd() *buildCmd {
 	buildCmd := &buildCmd{
 		Command: &cobra.Command{
-			Use: "build -t=<image_tag> [flags] <context_path>",
+			Use:                   "build -t=<image_tag> [flags] <context_path>",
 			DisableFlagsInUseLine: true,
-			Short: "Build docker image, optionally push to registries and/or load into docker daemon",
+			Short:                 "Build docker image, optionally push to registries and/or load into docker daemon",
 		},
 	}
 	buildCmd.Args = func(cmd *cobra.Command, args []string) error {
@@ -222,7 +222,7 @@ func (cmd *buildCmd) Build(contextDir string) error {
 	if err != nil {
 		return fmt.Errorf("failed to init image store: %s", err)
 	}
-	buildContext, err := context.NewBuildContext("/", contextDirAbs, imageStore)
+	buildContext, err := context.NewBuildContext("/", contextDirAbs, imageStore, "")
 	if err != nil {
 		return fmt.Errorf("failed to create initial build context: %s", err)
 	}
