+++ lib/docker/image/digest_test.go
@@ -42,7 +42,7 @@ func TestEmptyDigest(t *testing.T) {
 	defer os.RemoveAll(tmpDir)
 
 	targetPath := tmpDir + ".tar"
-	err = shell.ExecCommand(log.Infof, log.Errorf, "", "tar", "cvf", targetPath, "--files-from", "/dev/null")
+	err = shell.ExecCommand(log.Infof, log.Errorf, "", "", "tar", "cvf", targetPath, "--files-from", "/dev/null")
 	require.NoError(err)
 	defer os.Remove(targetPath)
