+++ lib/tario/testutils_test.go
@@ -21,5 +21,5 @@ import (
 
 func untarHelper(t, out string) error {
 	// Use p option to preserve permissions.
-	return shell.ExecCommand(log.Infof, log.Errorf, "", "tar", "xfp", t, "-C", out)
+	return shell.ExecCommand(log.Infof, log.Errorf, "", "", "tar", "xfp", t, "-C", out)
 }
